requirements.txt
# --- Core ---
python-dotenv>=1.0.0

# --- GCP / BigQuery ---
google-cloud-bigquery>=3.11.0
google-auth>=2.22.0
db-dtypes>=1.1.1

# --- Market Data ---
yfinance>=0.2.36

# --- Data Processing ---
pandas>=2.0.0
numpy>=1.24.0

# --- Statistical Modeling ---
scipy>=1.11.0
statsmodels>=0.14.0
arch>=6.2.0

# --- Visualization & Dashboard ---
plotly>=5.18.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
